refactor: extract bug-reports router (PR-2, Phase 1)
Pure move of the 6 bug-report endpoints out of server/routes.ts into a
new server/routes/bugReports.ts. Handlers moved verbatim (log lines,
status codes, response shapes unchanged). Router uses full-path routes
with per-route requireClerkUser/requireAdmin middleware and is mounted
bare via app.use(bugReportsRouter) at the exact position the first
bug-report endpoint previously occupied, preserving global registration
order.

Removed now-unused imports from routes.ts (confirmed unused by manual
audit; tsc has no noUnusedLocals): the bugReportArchival helpers, the
BugReport* contract schemas, and the unused Song import noted in the plan.

route-manifest snapshot unchanged (byte-identical), tsc green, 545/545
tests pass.
